+++ lib/src/classes.dart
@@ -4,8 +4,8 @@ class OSMdata {
   final double longitude;
   OSMdata(
       {required this.displayname,
-        required this.latitude,
-        required this.longitude});
+      required this.latitude,
+      required this.longitude});
   @override
   String toString() {
     return '$displayname, $latitude, $longitude';
@@ -32,7 +32,39 @@ class LatLong {
 class PickedData {
   final LatLong latLong;
   final String address;
-  final Map<String, dynamic> addressData;
+  final Address addressData;
 
   PickedData(this.latLong, this.address, this.addressData);
+}
+
+class Address {
+  String? houseNumber;
+  String? neighbourhood;
+  String? city;
+  String? state;
+  String? iSO31662Lvl4;
+  String? postcode;
+  String? country;
+  String? countryCode;
+
+  Address(
+      {this.houseNumber,
+      this.neighbourhood,
+      this.city,
+      this.state,
+      this.iSO31662Lvl4,
+      this.postcode,
+      this.country,
+      this.countryCode});
+
+  Address.fromJson(Map<String, dynamic> json) {
+    houseNumber = json['house_number'];
+    neighbourhood = json['neighbourhood'];
+    city = json['city'];
+    state = json['state'];
+    iSO31662Lvl4 = json['ISO3166-2-lvl4'];
+    postcode = json['postcode'];
+    country = json['country'];
+    countryCode = json['country_code'];
+  }
 }